Move cors-headers into middleware
