+++ src/server/index.js
@@ -9,7 +9,7 @@ exports.createServer = function(config){
   var app = express();
   // app.use(compress());
 
-  app.use(require('./cors-headers.js'));
+  app.use(require('./middleware/cors-headers.js'));
 
   if (config.sslKey && config.sslCert) {
     var options = {
